Move class ForumPosterComponent from package 'posted.game.pbem' -> 'posted.game'
